Split Data.cpp/.hpp files into one file for each subclass
