+++ rtlib/src/Data.cpp
@@ -8,7 +8,7 @@
 #include "rt/io/OBJReader.hpp"
 
 namespace et = envitools;
-using namespace Data;
+using namespace rt;
 
 // IO
 using OCVBridge = itk::OpenCVImageBridge;
@@ -28,41 +28,5 @@ Data::Pointer Data::Load(const fs::path& path)
             return std::make_shared<ENVI>(path);
         case "mesh":
             return std::make_shared<Mesh>(path);
-    }
-}
-
-Data::Image(const fs::path& path) { img_ = cv::imread(path.string()); }
-
-Data::ENVI(const fs::path& path)
-{
-    //Read in ENVI file from the file path, set the object to the envi_ private member
-    envi_ = et::ENVI envi(path); //or does it need to be envitools::ENVI envi(path);
-    //or is it et::ENVI envi_(path);
-
-}
-
-cv::Mat Data::ENVI::getImage(int idx)
-{
-    //Get image of particular spectral band
-    cv::Mat img = envi_.getBand(idx);
-
-    //envi_.closeFile();
-
-    return img;
-}
-
-Data::Mesh(const fs::path& path)
-{
-    //Read in mesh along with its texture
-    io::OBJReader reader;
-    reader.setPath(path);
-
-    try {
-        origMesh_ = reader.read();
-        cvFixedImage_ = reader.getTextureMat();
-        fixedImage_ = OCVBridge::CVMatToITKImage<Image8UC3>(cvFixedImage_);
-    } catch (const std::exception& e) {
-        std::cerr << e.what() << std::endl;
-        return EXIT_FAILURE;
     }
 }
